Added LEGAL_CHARS validator, checking that all characters in the value are acceptable.
Added INTERNET_MEDIA_TYPE validator, checking that the mime-type is legal according to the standard. Can be improved by adding extensive lists, but
that will prompt extra work in the future to maintain.

Added RTMP_URL validator, checking that the URL is an RTMP url.

MAX_LENGTH and MIN_LENGTH now checks using the default charset which is utf-8.



git-svn-id: https://svn.open.seria.net/svn/platform/trunk/seria@3524 d23b63ab-52c4-41c5-9713-938aad2d84f3